DRTII-250 Refactor to traits and case classes
