+++ server/src/test/scala/services/crunch/deskrecs/RunnableDeskRecsSpec.scala
@@ -77,16 +77,12 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
   def newBuffer: Buffer = Buffer(Iterable())
   val flexDesks = false
 
-  def flightsToDeskRecs(airportConfig: AirportConfig, minutesToCrunch: Int): (FlightsWithSplits, MillisSinceEpoch) => DeskRecMinutes = {
-    Crunch.flightsToDeskRecs(minutesToCrunch, airportConfig, flexDesks, CrunchMocks.mockCrunch)
-  }
-
   "Given a RunnableDescRecs with a mock PortStateActor and mock crunch " +
     "When I give it a millisecond of 2019-01-01T00:00 " +
     "I should see a request for flights for 2019-01-01 00:00 to 00:30" >> {
     val portStateProbe = TestProbe("port-state")
     val mockPortStateActor = system.actorOf(Props(classOf[MockPortStateActor], portStateProbe, noDelay))
-    val (millisToCrunchSourceActor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minutesToCrunch, defaultAirportConfig, flightsToDeskRecs(defaultAirportConfig, minutesToCrunch), newBuffer).run()
+    val (millisToCrunchSourceActor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(defaultAirportConfig, minutesToCrunch, mockCrunch), newBuffer).run()
     val askableSource: AskableActorRef = millisToCrunchSourceActor
 
     val midnight20190101 = SDate("2019-01-01T00:00")
@@ -108,7 +104,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     val portStateProbe = TestProbe("port-state")
     val mockPortStateActor = system.actorOf(Props(classOf[MockPortStateActor], portStateProbe, longDelay))
 
-    val (millisToCrunchSourceActor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minutesToCrunch, defaultAirportConfig, flightsToDeskRecs(defaultAirportConfig, minutesToCrunch), newBuffer).run()
+    val (millisToCrunchSourceActor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(defaultAirportConfig, minutesToCrunch, mockCrunch), newBuffer).run()
     val askableSource: AskableActorRef = millisToCrunchSourceActor
 
     val midnight20190101 = SDate("2019-01-01T00:00")
@@ -145,7 +141,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     val procTimes: Map[Terminal, Map[PaxTypeAndQueue, Double]] = Map(T1 -> Map(eeaMachineReadableToDesk -> 30d / 60, visaNationalToDesk -> 60d / 60))
     val testAirportConfig = defaultAirportConfig.copy(terminalProcessingTimes = procTimes)
 
-    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minutesToCrunch, testAirportConfig, flightsToDeskRecs(testAirportConfig, minutesToCrunch), newBuffer).run()
+    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(testAirportConfig, minutesToCrunch, mockCrunch), newBuffer).run()
     val millisToCrunchSourceActor: AskableActorRef = actor
 
     val epoch = SDate(scheduled).millisSinceEpoch
@@ -183,7 +179,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     val procTimes: Map[Terminal, Map[PaxTypeAndQueue, Double]] = Map(T1 -> Map(eeaMachineReadableToDesk -> 30d / 60, visaNationalToDesk -> 60d / 60))
     val testAirportConfig = defaultAirportConfig.copy(terminalProcessingTimes = procTimes)
 
-    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minutesToCrunch, testAirportConfig, flightsToDeskRecs(testAirportConfig, minutesToCrunch), newBuffer).run()
+    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(testAirportConfig, minutesToCrunch, mockCrunch), newBuffer).run()
     val millisToCrunchSourceActor: AskableActorRef = actor
 
     val epoch = SDate(scheduled).millisSinceEpoch
@@ -223,7 +219,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     val procTimes: Map[Terminal, Map[PaxTypeAndQueue, Double]] = Map(T1 -> Map(eeaMachineReadableToDesk -> 30d / 60, visaNationalToDesk -> 60d / 60))
     val testAirportConfig = defaultAirportConfig.copy(terminalProcessingTimes = procTimes)
 
-    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minutesToCrunch, testAirportConfig, flightsToDeskRecs(testAirportConfig, minutesToCrunch), newBuffer).run()
+    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(testAirportConfig, minutesToCrunch, mockCrunch), newBuffer).run()
     val millisToCrunchSourceActor: AskableActorRef = actor
 
     val scheduledSd = SDate(scheduled)
@@ -266,7 +262,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     val procTimes: Map[Terminal, Map[PaxTypeAndQueue, Double]] = Map(T1 -> Map(eeaMachineReadableToDesk -> 30d / 60))
     val testAirportConfig = defaultAirportConfig.copy(terminalProcessingTimes = procTimes)
 
-    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minutesToCrunch, testAirportConfig, flightsToDeskRecs(testAirportConfig, minutesToCrunch), newBuffer).run()
+    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(testAirportConfig, minutesToCrunch, mockCrunch), newBuffer).run()
     val millisToCrunchSourceActor: AskableActorRef = actor
 
     val noonMillis = SDate(pcpOne).millisSinceEpoch
@@ -434,7 +430,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
       divertedQueues = Map(Queues.NonEeaDesk -> Queues.EeaDesk)
     )
 
-    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minutesToCrunch, testAirportConfig, flightsToDeskRecs(testAirportConfig, minutesToCrunch), newBuffer).run()
+    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(testAirportConfig, minutesToCrunch, mockCrunch), newBuffer).run()
     val millisToCrunchSourceActor: AskableActorRef = actor
 
     val scheduledMillis = SDate(scheduled).millisSinceEpoch
@@ -514,7 +510,7 @@ class RunnableDeskRecsSpec extends CrunchTestLike {
     mockPortStateActor ! SetFlights(flight)
 
     val minsInADay = 1440
-    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, minsInADay, defaultAirportConfig, flightsToDeskRecs(defaultAirportConfig, minsInADay), newBuffer).run()
+    val (actor: ActorRef, _) = RunnableDeskRecs(mockPortStateActor, FlexedPortDeskRecs(defaultAirportConfig, minsInADay, mockCrunch), newBuffer).run()
     val millisToCrunchSourceActor: AskableActorRef = actor
 
     val epoch = SDate(scheduled).millisSinceEpoch
